test(select): update use-select.ts l:261
